moved load_batches to data and made it a package
