+++ campolina/data/__init__.py
@@ -1,0 +1,2 @@
+from .bam_util import BamIndex
+from .load import load_batches
