+++ train.py
@@ -7,82 +7,11 @@ from tqdm import tqdm
 import torch
 from torch.optim import AdamW
 
-from campolina.data.bam_util import BamIndex
-from campolina.data.pod5_util import get_reads, process_chunk
+from campolina.data import BamIndex, load_batches
 from campolina.model.model import EventDetector
 from loss import CustomLoss
 
 #torch.manual_seed(12345)
-
-def load_batches(bam_idx: BamIndex, pod5_path: str, batch_size: int, predict: bool = False): 
-    """
-    Load data for training or validation.
-    """
-
-    # TODO what if i preload a batch of batch_size signals and then yield batches as long as i can and then load batch_size signal again?
-    current_batch = []; current_borders = []; current_identifiers = []
-
-    for read in get_reads(pod5_path):
-        for alignment in bam_idx.get_alignment(str(read.read_id)):
-            if alignment is None: continue
-
-            signal_chunks, chunk_borders, chunk_identifiers = process_chunk(
-                aln=alignment, 
-                read=read, 
-                predict=predict, 
-                adjust_type=None,
-            )
-
-            if signal_chunks is None:
-                tqdm.write(f'Could not extract info for read {read.read_id}')
-                continue
-
-            #tqdm.write(f'Signal {read.read_id} has {len(signal_chunks)}')
-
-            if len(current_batch) + len(signal_chunks) > batch_size:
-
-                to_take = batch_size - len(current_batch)
-                current_batch.extend(signal_chunks[:to_take])
-                current_borders.extend(chunk_borders[:to_take])
-
-                if predict:
-                    current_identifiers.extend(chunk_identifiers[:to_take])
-                    yield np.array(current_batch), np.array(current_borders), np.array(current_identifiers)
-                else:
-                    #print(read.read_id)
-                    yield np.array(current_batch), np.array(current_borders)
-
-                remaining = len(signal_chunks) - to_take
-
-                while remaining >= batch_size:
-                    current_batch = signal_chunks[to_take:to_take+batch_size]
-                    current_borders = chunk_borders[to_take:to_take+batch_size]
-
-                    if predict:
-                        current_identifiers = chunk_identifiers[to_take:to_take+batch_size]
-                        yield np.array(current_batch), np.array(current_borders), np.array(current_identifiers)
-                    else:
-                        #print(read.read_id)
-                        yield np.array(current_batch), np.array(current_borders)
-
-                    to_take = to_take + batch_size
-                    remaining = remaining - batch_size
-
-                current_batch = signal_chunks[to_take:]
-                current_borders = chunk_borders[to_take:]
-
-                if predict: current_identifiers = chunk_identifiers[to_take:]
-
-            else:
-                current_batch.extend(signal_chunks)
-                current_borders.extend(chunk_borders)
-                if predict: current_identifiers.extend(chunk_identifiers)
-
-    if predict:
-        yield np.array(current_batch), np.array(current_borders), np.array(current_identifiers)
-    else:
-        #print(read.read_id)
-        yield np.array(current_batch), np.array(current_borders)
 
 def test_model(bam_idx: BamIndex, model: EventDetector, device, loss_f, scope: dict, valid: bool = False):
     model.eval()
